+++ src/clusterfock/cc/ccsd.py
@@ -36,7 +36,7 @@ class GCCSD(CoupledCluster):
             o=basis.o,
         )
 
-        rhs = CoupledClusterParameter(t.orders, t.N, t.M)
+        rhs = CoupledClusterParameter(t.orders, t.N, t.M, dtype=t.dtype)
         rhs.initialize_dicts({1: rhs1, 2: rhs2})
 
         return rhs
@@ -57,7 +57,7 @@ class GCCSD(CoupledCluster):
             o=basis.o,
         )
 
-        rhs = CoupledClusterParameter(l.orders, l.N, l.M)
+        rhs = CoupledClusterParameter(l.orders, l.N, l.M, dtype=l.dtype)
         rhs.initialize_dicts({1: rhs1, 2: rhs2})
 
         return rhs
